scripts: Download Linux binaries on Windows.
This also includes refactoring so the download bits don't just grab
functions out of a random script.
